# backend/openapi.json
{
  "components": {
    "schemas": {
      "CurrencyRead": {
        "properties": {
          "code": {
            "pattern": "^[A-Z]{3}$",
            "title": "Code",
            "type": "string"
          },
          "exponent": {
            "title": "Exponent",
            "type": "integer"
          }
        },
        "required": [
          "code",
          "exponent"
        ],
        "title": "CurrencyRead",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Body of every error response except 422, for `responses={404: {\"model\": ErrorResponse}}`.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "SessionCreate": {
        "additionalProperties": false,
        "description": "Body of `POST /auth/session`.",
        "properties": {
          "id_token": {
            "maxLength": 8192,
            "minLength": 1,
            "title": "Id Token",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id_token",
          "timezone"
        ],
        "title": "SessionCreate",
        "type": "object"
      },
      "SpaceRead": {
        "properties": {
          "currency": {
            "$ref": "#/components/schemas/CurrencyRead"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "type": {
            "enum": [
              "personal",
              "household"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "currency",
          "timezone"
        ],
        "title": "SpaceRead",
        "type": "object"
      },
      "UserRead": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "display_name"
        ],
        "title": "UserRead",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Personal Finance App",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/auth/session": {
      "delete": {
        "description": "Log out: clear the cookie. Always 204, even if there was no session.",
        "operationId": "delete_session_auth_session_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          }
        },
        "summary": "Delete Session",
        "tags": [
          "auth"
        ]
      },
      "post": {
        "description": "Exchange a Firebase ID token (signed in less than 5 minutes ago) for the session cookie.\n\nThe cookie is HttpOnly and lasts 14 days. The first login also creates the user and their\npersonal space \"Mi espacio\" (UYU, in the given timezone).",
        "operationId": "create_session_auth_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "auth"
        ]
      }
    },
    "/healthz": {
      "get": {
        "description": "Liveness probe: no database or external dependencies.",
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Healthz",
        "tags": [
          "health"
        ]
      }
    },
    "/me": {
      "get": {
        "description": "The logged-in user.",
        "operationId": "read_me_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "summary": "Read Me",
        "tags": [
          "users"
        ]
      }
    },
    "/spaces": {
      "get": {
        "description": "The spaces where the user is a member.",
        "operationId": "list_spaces_spaces_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SpaceRead"
                  },
                  "title": "Response List Spaces Spaces Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "summary": "List Spaces",
        "tags": [
          "spaces"
        ]
      }
    },
    "/spaces/{space_id}": {
      "get": {
        "description": "A space the user belongs to, with its currency and ISO 4217 exponent.",
        "operationId": "read_space_spaces__space_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Space",
        "tags": [
          "spaces"
        ]
      }
    }
  }
}